feat: Add camelizeKeysPlugin
