--- src/environment.d.ts
@@ -1,18 +1,0 @@
-declare global {
-  namespace NodeJS {
-    interface ProcessEnv {
-      NODE_ENV: 'development' | 'production';
-      PORT?: string;
-      MONGODB_USER: string;
-      MONGODB_PASSWORD: string;
-      MONGODB_DATABASE: string;
-      MONGODB_CONNECT_STRING: string;
-      JWT_SECRET: string;
-      LOG_FILE_DIR?: string;
-    }
-  }
-}
-
-// If this file has no import/export statements (i.e. is a script)
-// convert it into a module by adding an empty export statement.
-export {};
